use max_length and server side setting
