+++ docs/v3/develop/settings-ref.mdx
@@ -71,13 +71,6 @@ If True, enables debug mode which may provide additional logging and debugging f
 
 **TOML dotted key path**: `deployments`
 
-### `events`
-Settings for controlling events behavior
-
-**Type**: [EventsSettings](#eventssettings)
-
-**TOML dotted key path**: `events`
-
 ### `experiments`
 Settings for controlling experimental features
 
@@ -446,21 +439,6 @@ The default Docker namespace to use when building images.
 
 **Supported environment variables**:
 `PREFECT_DEPLOYMENTS_DEFAULT_DOCKER_BUILD_NAMESPACE`, `PREFECT_DEFAULT_DOCKER_BUILD_NAMESPACE`
-
----
-## EventsSettings
-Settings for controlling behavior of the events subsystem
-### `maximum_event_name_length`
-The maximum length of an event name.
-
-**Type**: `integer`
-
-**Default**: `1024`
-
-**TOML dotted key path**: `events.maximum_event_name_length`
-
-**Supported environment variables**:
-`PREFECT_EVENTS_MAXIMUM_EVENT_NAME_LENGTH`
 
 ---
 ## ExperimentsSettings
@@ -1375,6 +1353,18 @@ Which cache implementation to use for the events system. Should point to a modul
 
 **Supported environment variables**:
 `PREFECT_SERVER_EVENTS_MESSAGING_CACHE`, `PREFECT_MESSAGING_CACHE`
+
+### `maximum_event_name_length`
+The maximum length of an event name.
+
+**Type**: `integer`
+
+**Default**: `1024`
+
+**TOML dotted key path**: `server.events.maximum_event_name_length`
+
+**Supported environment variables**:
+`PREFECT_SERVER_EVENTS_MAXIMUM_EVENT_NAME_LENGTH`, `PREFECT_EVENTS_MAXIMUM_EVENT_NAME_LENGTH`
 
 ---
 ## ServerFlowRunGraphSettings
