+++ schemas/settings.schema.json
@@ -306,22 +306,6 @@
             "title": "DeploymentsSettings",
             "type": "object"
         },
-        "EventsSettings": {
-            "description": "Settings for controlling behavior of the events subsystem",
-            "properties": {
-                "maximum_event_name_length": {
-                    "default": 1024,
-                    "description": "The maximum length of an event name.",
-                    "supported_environment_variables": [
-                        "PREFECT_EVENTS_MAXIMUM_EVENT_NAME_LENGTH"
-                    ],
-                    "title": "Maximum Event Name Length",
-                    "type": "integer"
-                }
-            },
-            "title": "EventsSettings",
-            "type": "object"
-        },
         "ExperimentsSettings": {
             "description": "Settings for configuring experimental features",
             "properties": {
@@ -1199,6 +1183,17 @@
                     ],
                     "title": "Messaging Cache",
                     "type": "string"
+                },
+                "maximum_event_name_length": {
+                    "default": 1024,
+                    "description": "The maximum length of an event name.",
+                    "exclusiveMinimum": 0,
+                    "supported_environment_variables": [
+                        "PREFECT_SERVER_EVENTS_MAXIMUM_EVENT_NAME_LENGTH",
+                        "PREFECT_EVENTS_MAXIMUM_EVENT_NAME_LENGTH"
+                    ],
+                    "title": "Maximum Event Name Length",
+                    "type": "integer"
                 }
             },
             "title": "ServerEventsSettings",
@@ -2187,11 +2182,6 @@
         },
         "deployments": {
             "$ref": "#/$defs/DeploymentsSettings",
-            "supported_environment_variables": []
-        },
-        "events": {
-            "$ref": "#/$defs/EventsSettings",
-            "description": "Settings for controlling events behavior",
             "supported_environment_variables": []
         },
         "experiments": {
