+++ tests/events/server/test_events_schema.py
@@ -295,3 +295,8 @@ def test_finding_resources_in_role(example_event: Event):
     ]
     assert [r.id for r in example_event.resources_in_role["role-2"]] == ["related-3"]
     assert example_event.resources_in_role["role-3"] == []
+
+
+def test_event_name_length(example_event: Event):
+    with pytest.raises(ValidationError, match="Event name must be at most"):
+        Event(event="x" * 1025, **example_event.model_dump(exclude={"event"}))
