+++ tests/test_settings.py
@@ -212,7 +212,6 @@ SUPPORTED_SETTINGS = {
         "test_value": timedelta(seconds=60),
         "legacy": True,
     },
-    "PREFECT_EVENTS_MAXIMUM_EVENT_NAME_LENGTH": {"test_value": 1024},
     "PREFECT_EVENTS_MAXIMUM_LABELS_PER_RESOURCE": {"test_value": 10, "legacy": True},
     "PREFECT_EVENTS_MAXIMUM_RELATED_RESOURCES": {"test_value": 10, "legacy": True},
     "PREFECT_EVENTS_MAXIMUM_SIZE_BYTES": {"test_value": 10, "legacy": True},
@@ -314,6 +313,7 @@ SUPPORTED_SETTINGS = {
     "PREFECT_SERVER_EVENTS_EXPIRED_BUCKET_BUFFER": {
         "test_value": timedelta(seconds=60)
     },
+    "PREFECT_SERVER_EVENTS_MAXIMUM_EVENT_NAME_LENGTH": {"test_value": 1024},
     "PREFECT_SERVER_EVENTS_MAXIMUM_LABELS_PER_RESOURCE": {"test_value": 10},
     "PREFECT_SERVER_EVENTS_MAXIMUM_RELATED_RESOURCES": {"test_value": 10},
     "PREFECT_SERVER_EVENTS_MAXIMUM_SIZE_BYTES": {"test_value": 10},
